+++ apps/weather/src/constants/enums/weather.ts
