+++ apps/weather/src/constants/types/city.ts
