+++ apps/weather/src/constants/types/weather.ts
